tc: add bytecode dump for eBPF with tc monitor
So far, `tc monitor` would only dump bytecode for cBPF programs that it
reads from netlink -- the kernel would not send bytecode for eBPF.

Lately, bytecode transmission over the netlink socket has been added to
the kernel for eBPF as well. `tc monitor` dumps it already, but as cBPF
code. This commit is a fix to print eBPF bytecode in eBPF format.

Signed-off-by: Quentin Monnet <quentin.monnet@6wind.com>